Stop loading packages async
Loading packages async was complicated and made compulation of staleness very complicated.

Importantly, as Package.Stale is dependant on the state of each of its dependant packages
it cannot be computed til the whole dependency graph is computed; so there is no advantage to
trying to load parts of the package graph concurrently.

Simplify the code by removing async package loading. Packages are now loaded head of time
into a Context then Test or Build actions are performed on that context.