gb-vendor: add support for fetching a tag
Fixes #209

This PR adds support for fetching from a tag for all repos, however
tag is ignored by all but git. Hopefully someone with hg or bzr
skills can improve this.

This PR also adds a check to ensure that if a repo, again git, is fetched
from a tag or revision, that update will not let the caller update, as
there is no path from one revision to the next if you are in a detached
state.

The error message is very bad, but at least we don't promise something that
isn't possible.

% gb vendor fetch -tag r2015.06.03 gopkg.in/mgo.v2
% gb vendor update gopkg.in/mgo.v2
FATAL command "update" failed: cannot checkout HEAD
FATAL command "vendor" failed: exit status 1