cmd/gb: add experimental depset command
depset is an experimental command to explore analysing dependency sets.
It produces output like this, http://paste.ubuntu.com/11739677/

At the moment it has worse than quadratic behaviour, making it unsuitable for anything serious.

The goal of depset is to develop a set of functions in the `cmd` package that can be used by tools like `gb-vendor` to analyse dependencies as a set rather than as individual imports.
Dependency sets are groups of packages that are related by being children of a single source root. A canonical example of a source root is $GOPATH/src, another is $PROJECT/src. However we can be smarter and create sets for each vendored import tree, which in turn allows us to determine which dependencies are no longer referenced with more accuracy than the current gb vendor purge command.

Another function of dep sets is to determine missing packages, the output of which could be fed back into gb vendor fetch to drive recursive depdendency fetching.