cmd/gb-vendor: do not apply tag/branch/revision during recursive fetch
Fixes #270

If gb vendor fetch causes another missing depdendency to be exposed (this is a
fancy way of saying, fetch recursively), then do not apply and tag, branch, or
revision values as it is extremely unlikely they will make sense in the context
of a transitive dependency.

This does mean that a dependency discovered during recursive fetching will be fetched
from the head of it's master branch. If this is not the desired behaviour, then
callers should use `-no-recurse` and fetch each dependency individually.