Introduce ctxString
It's a horrible name, I couldn't think of a better one.

ctxString represents a hyphen seperated string of identifying
properties of the Context. GOOS and GOARCH are the key ones, but
others like -tags, -race, -nocgo, etc may be present.

Not supported yet, but if present, additional context elements must
be sorted aphabetically.