Avoid passing build flags to test process.
The gb test command advertises support for all valid
flags to gb build (e.g. -P), but the logic which determines which flags
should be sent to the test binary was not aware of them and blithely
passed them through. Add all valid build flags to the map indicating
they are meant for gb and not to passed through.