Do not force every package to depend on runtime explicitly
Fixes #323

Forcing every package to have an explicit dependency on the runtime package,
even though 99% of the time they would have a transitive dependency was lame.
It also added many edges to the action graph that were not necessary.

However, the _implicit_ runtime dependency is actually a _link time_ depdendency
not a compile time dependency; pkg runtime must be up to date for the link phase
to run, compilation without linking is not affected.

Realising this we can add the runtime as a depdendency of all _main_ packages,
which reduces the number of redundant edges signficantly.

This also makes it possible to add additional link time depdendencies later, like
math, or runtime/race as required.