new package: gb/test
The goal of not placing the test code directly in gb/ was to demonstrate
that testing did not require tight coupling with gb/ itself -- and this is
true, placing the test code outside gb/ has demonstrated that testing is
a unique function.

However, placing test code inside gb/cmd was a poor choice, and the lesser
evil would be to create yet another package and place the test support code
in that, which is what this PR does.

In the future several things will probably happen.

1. The test code may be integrated into gb/ proper, this would probably be a
net win for the surface area of the greater gb library.

2. The code in gb/cmd will be reduced or eliminated.