cmd/gb: fix parsing of passToAll test arguments
Fixes #473

Handling of flags that are parsed by both gb and a sub process is tricky.
So that the arguments are not parsed as package parameters by gb test
we move them to the front of the argument list when parsed by gb, but
leave them at the end of the list for the test binary to parse after
invocation.