Refactor package stale computation
Previously staleness of a package was handled with a large map of packages
that were part of the standard library. This was suboptimal, and a bit gross.

In a pre cross compilation, pre build tag, pre race detection world, it was
roughly safe to assume that the stdlib was always up to date. Now, with all
these features in play, we may need to compile the standard lib and cache
those files in the project's cache.

This change removes the hard coded map, and bases its logic on the
build.Package definition and Context parameters.

This also meant implementing vendor/ support for the stdlib. vendor/ is not
supported outside the stdlib, gb will ignore vendor/ paths. See
https://getgb.io/faq/#govendorexperiment