Refactor pkgfile to become pkgpath
pkgpath always returns the location where a previously compiled .a for this
pacakge would be found. pkgpath considers elements of the context like tags
cross compilation and race status.

pkgpath is used exclusively by isStale to determine if the package is up to date.
However, pkgpath is _not_ used to compute the target to install a .a file. This
discontinuity takes into account the fact that if the stdlib is missing for a
particular build context, replacement files should be written to installpath,
the $PROJECT/pkg/ directory.