Always assume the stdlib is up to date if present
Fixes #515

In the 0.3.x series we removed the hard coded list of packages to assume
are up to date and let that fall through to `isStale`. This worked ok, except
the Go installer for Darwin is broken and its timestamps are not correct.

Work around this by assuming that _iff_ the package is the standard lib and
_iff_ we are not cross compiling, then the standard lib is up to date. This
mimicks the pre 0.3.x behaviour and avoids the problem with broken timestamps
in the darwin binary installer.