Memoize Package.Imports
Previously gb.Package.Imports was a method which would fetch packages'
from the context package cache matching the import paths of
gb.Package.Package.Imports.

This change causes the fetch to happen earlier, which will avoid the
panic happening before the build starts. A followup change can now
change the panic to be an error.