Move NewPackage to be a method on Context.
Move the NewPackage helper to gb.Context. As the first parameter was
always a concrete type, not an interface, this makes sense. Making
this a method on Context means that types which embed a Context, like
Package get access to it, allowing a caller who has a *Package to create
another *Package with the same underlying context.