Refactor importers to not be a slice, but a stack.
When resolving packages they should be done so in an order that matches
the expected search order. First we look in GOROOT, if that misses, then
we look in $PROJECT/src, then if that fails we look in
$PROJECT/vendor/src.

Currently each of these importers has slightly different behaviour, that
will be fixed in a future PR. This current PR makes it possible to slip
another importer, provided by the depfile below the current ones, at the
lowest level.