cmd/gb-vendor: search $GOROOT/src/vendor when resolving recursive deps
Fixes #635

golang/go#16333 renamed $GOROOT/src/vendor/golang.org to
$GOROOT/src/vendor/golang_org. This solved a nasty problem where code in
GOROOT would shadow any other copy of the net/http2 library, but meant
that gb vendor could no longer resolve the package and try to fetch it
(even if that was pointless).

To solve this, include $GOROOT/src/vendor in the set of search paths so
that golang_org/net/http/... is always found (if present).

I'll add a test for this in the integration test repo.