Run compiler from the project directory
In an attempt to solve issue #406, run the compiler instead of from
the package directory from the root of the project. Make all paths
passed to the compiler relative to the project root.

This appears to work, and the tests still pass; but it corrects the
output of compiler errors and they now appear with filenames relative
to the `GB_PROJECT_DIR`.